+++ week9/backend/print_vector.hpp
@@ -5,7 +5,7 @@
 #include <iostream>
 
 void print_vector(std::vector<int>& vec){
-	for (unsigned int i = 0; i < vec.size(); i++)
+	for (size_t i = 0; i < vec.size(); i++)
 		std::cout<< vec[i] << " ";
 	std::cout << std::endl;
 }
