+++ week9/sorting_methods/bubbleSort.hpp
@@ -12,9 +12,9 @@ sortData bubbleSort(std::vector<int>& avector, char type, bool print){
     unsigned long long array_accesses = 0;
     unsigned long long comparisons = 0;
     //---------------------------------------------------------------------------------
-    for (int pass = avector.size()-1; pass > 0; pass -= 1){
+    for (size_t pass = avector.size()-1; pass > 0; pass -= 1){
         sorted = true;
-		for (int i = 0; i < pass; i++){
+		for (size_t i = 0; i < pass; i++){
             comparisons ++ ;
             if (print)
                 printTowers(avector);
