+++ week9/sorting_methods/reverseBubbleSort.hpp
@@ -9,9 +9,9 @@
 sortData reverseBubbleSort(std::vector<int>& avector, char type){
     bool sorted = false;
     int swaps = 0;
-    for (int pass = avector.size()-1; pass > 0; pass -= 1){
+    for (size_t pass = avector.size()-1; pass > 0; pass -= 1){
         sorted = true;
-		for (int i = 0; i < pass; i++){
+		for (size_t i = 0; i < pass; i++){
 			if (avector[i] < avector[i+1]){
                 swap(avector, i, i+1);
                 swaps ++ ;
