+++ week9/sorting_methods/shellSortCustomGap.hpp
@@ -9,9 +9,9 @@
 sortData shellSortCustomGaps(std::vector<int>& vector, std::vector<int>& gaps, char type, bool print){
     unsigned long long array_accesses = 0;
     unsigned long long comparisons = 0;
-	for (int i = 0 ; i < gaps.size(); i ++ ){
-        int gap = gaps[i];
-        for (int start = 0; start < gap; start++) {
+	for (size_t i = 0 ; i < gaps.size(); i ++ ){
+        size_t gap = gaps[i];
+        for (size_t start = 0; start < gap; start++) {
             sortData s = gapInsertionSort(vector, start, gap, print);
 			array_accesses += s.array_accesses;
 			comparisons += s.comparisons;
